fix(native): stop yanking the transcript while streaming
The follow-tail pin treated a 240px near-bottom band as still following,
so scrolling up during a turn was overwritten on the next token (#724).
Pin only while the reader is at the tail; scrolling away pauses follow.

While a turn is running, Enter queues the next prompt instead of no-op,
and the composer stays usable. Queued lines drain in order when the
turn ends.
